refactor(git-utils) update imports
